demo works, and now it is in utils
